DeleteOriginTest fix attempt for non-existing origin directory
BUG=61676
TEST=SandboxQuotaClientTest.DeleteOriginTest
TBR=tzik@google.com

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@85598 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 08ebe0a22ebf0195069000b143453a4ef7ee3859

+++ webkit/fileapi/file_system_context.cc
@@ -77,6 +77,8 @@ bool FileSystemContext::DeleteDataForOriginOnFileThread(
 
   FilePath path_for_origin =
       sandbox_provider()->GetBaseDirectoryForOrigin(origin_url);
+  if (!file_util::PathExists(path_for_origin))
+    return true;
   return file_util::Delete(path_for_origin, true /* recursive */);
 }
 
@@ -87,6 +89,8 @@ bool FileSystemContext::DeleteDataForOriginAndTypeOnFileThread(
 
   FilePath path_for_origin =
       sandbox_provider()->GetBaseDirectoryForOriginAndType(origin_url, type);
+  if (!file_util::PathExists(path_for_origin))
+    return true;
   return file_util::Delete(path_for_origin, true /* recursive */);
 }
